+++ database/migrations/2026_01_22_010825_create_proyecto_avances_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('proyecto_id')->constrained()->cascadeOnDelete();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->date('fecha');
-            $table->decimal('porcentaje_avance', 5, 2);
+            $table->unsignedTinyInteger('porcentaje_avance');
             $table->text('comentario')->nullable();
             $table->timestamps();
         });
